refactor: move common editor code to BaseEditorClass
