+++ packages/common/src/editors/inputEditor.ts
@@ -1,19 +1,9 @@
-import { BindingEventService } from '@slickgrid-universal/binding';
 import { createDomElement, setDeepValue, toSentenceCase } from '@slickgrid-universal/utils';
-import { SlickEventData, type SlickGrid } from '../core/index.js';
+import { SlickEventData } from '../core/index.js';
 import { textValidator } from '../editorValidators/textValidator.js';
-import type {
-  Column,
-  ColumnEditor,
-  CompositeEditorOption,
-  Editor,
-  EditorArguments,
-  EditorValidationResult,
-  EditorValidator,
-  GridOption,
-  ValidateOption,
-} from '../interfaces/index.js';
+import type { CompositeEditorOption, Editor, EditorArguments, EditorValidationResult, ValidateOption } from '../interfaces/index.js';
 import { getDescendantProperty } from '../services/utilities.js';
+import { BaseEditorClass } from './baseEditorClass.js';
 
 const DEFAULT_DECIMAL_PLACES = 0;
 
@@ -21,8 +11,7 @@ const DEFAULT_DECIMAL_PLACES = 0;
  * An example of a 'detached' editor.
  * KeyDown events are also handled to provide handling for Tab, Shift-Tab, Esc and Ctrl-Enter.
  */
-export class InputEditor implements Editor {
-  protected _bindEventService: BindingEventService;
+export class InputEditor extends BaseEditorClass implements Editor {
   protected _isInternalFocusInProgress = false;
   protected _input!: HTMLInputElement | undefined;
   protected _inputType = 'text';
@@ -31,48 +20,18 @@ export class InputEditor implements Editor {
   protected _originalValue?: number | string;
   protected _timer?: any;
 
-  /** is the Editor disabled? */
-  disabled = false;
-
-  /** SlickGrid Grid object */
-  grid: SlickGrid;
-
-  /** Grid options */
-  gridOptions: GridOption;
-
   constructor(
     protected readonly args: EditorArguments,
     inputType = 'text'
   ) {
-    this.grid = args.grid;
-    this.gridOptions = args.grid?.getOptions() as GridOption;
-    this._bindEventService = new BindingEventService();
+    super(args);
     this.inputType = inputType;
     this.init();
-  }
-
-  /** Get Column Definition object */
-  get columnDef(): Column {
-    return this.args.column;
-  }
-
-  /** Get Column Editor object */
-  get columnEditor(): ColumnEditor {
-    return this.columnDef?.editor || ({} as ColumnEditor);
-  }
-
-  /** Getter for the item data context object */
-  get dataContext(): any {
-    return this.args.item;
   }
 
   /** Getter for the Editor DOM Element */
   get editorDomElement(): any {
     return this._input;
-  }
-
-  get hasAutoCommitEdit(): boolean {
-    return this.gridOptions.autoCommitEdit ?? false;
   }
 
   /** Getter of input type (text, number, password) */
@@ -83,11 +42,6 @@ export class InputEditor implements Editor {
   /** Setter of input type (text, number, password) */
   set inputType(type: string) {
     this._inputType = type;
-  }
-
-  /** Get the Validator function, can be passed in Editor property or Column Definition */
-  get validator(): EditorValidator | undefined {
-    return this.columnEditor?.validator ?? this.columnDef?.validator;
   }
 
   init(): void {
@@ -345,17 +299,7 @@ export class InputEditor implements Editor {
 
   /** when it's a Composite Editor, we'll check if the Editor is editable (by checking onBeforeEditCell) and if not Editable we'll disable the Editor */
   protected applyInputUsabilityState(): void {
-    const activeCell = this.grid.getActiveCell();
-    const isCellEditable = this.grid.onBeforeEditCell
-      .notify({
-        ...activeCell,
-        item: this.dataContext,
-        column: this.args.column,
-        grid: this.grid,
-        target: 'composite',
-        compositeEditorOptions: this.args.compositeEditorOptions,
-      })
-      .getReturnValue();
+    const isCellEditable = super.checkInputUsabilityState();
     this.disable(isCellEditable === false);
   }
