+++ packages/common/src/editors/selectEditor.ts
@@ -8,30 +8,27 @@ import {
   sortCollectionWithOptions,
 } from '../commonEditorFilter/commonEditorFilterUtils.js';
 import { Constants } from '../constants.js';
-import { SlickEventData, type SlickGrid } from '../core/index.js';
+import { SlickEventData } from '../core/index.js';
 import { buildMsSelectCollectionList, CollectionService, findOrDefault, type TranslaterService } from '../services/index.js';
 import { getDescendantProperty, getTranslationPrefix } from '../services/utilities.js';
 import type {
   CollectionCustomStructure,
   CollectionOption,
   CollectionOverrideArgs,
-  Column,
-  ColumnEditor,
   CompositeEditorOption,
   Editor,
   EditorArguments,
   EditorValidationResult,
-  EditorValidator,
-  GridOption,
   Locale,
   SelectOption,
   ValidateOption,
 } from './../interfaces/index.js';
+import { BaseEditorClass } from './baseEditorClass.js';
 
 /**
  * Slickgrid editor class for multiple/single select lists
  */
-export class SelectEditor implements Editor {
+export class SelectEditor extends BaseEditorClass implements Editor {
   protected _isValueTouched = false;
 
   /** Locales */
@@ -51,9 +48,6 @@ export class SelectEditor implements Editor {
 
   /** Editor DOM element */
   editorElm?: HTMLElement;
-
-  /** is the Editor disabled? */
-  disabled = false;
 
   /** Editor Multiple-Select options */
   editorElmOptions!: Partial<MultipleSelectOption>;
@@ -82,14 +76,8 @@ export class SelectEditor implements Editor {
   /** The property name for values in the collection */
   valueName!: string;
 
-  /** Grid options */
-  gridOptions: GridOption;
-
   /** Do we translate the label? */
   enableTranslateLabel = false;
-
-  /** SlickGrid Grid object */
-  grid: SlickGrid;
 
   /** Final collection displayed in the UI, that is after processing filter/sort/override */
   finalCollection: any[] = [];
@@ -98,8 +86,7 @@ export class SelectEditor implements Editor {
     protected readonly args: EditorArguments,
     protected readonly isMultipleSelect: boolean
   ) {
-    this.grid = args.grid;
-    this.gridOptions = (this.grid.getOptions() || {}) as GridOption;
+    super(args);
     if (this.gridOptions?.translater) {
       this._translaterService = this.gridOptions.translater;
     }
@@ -185,23 +172,8 @@ export class SelectEditor implements Editor {
     return this.columnEditor?.collectionOptions;
   }
 
-  /** Get Column Definition object */
-  get columnDef(): Column {
-    return this.args.column;
-  }
-
   get columnId(): string | number {
     return this.columnDef?.id ?? '';
-  }
-
-  /** Get Column Editor object */
-  get columnEditor(): ColumnEditor {
-    return (this.columnDef?.editor ?? {}) as ColumnEditor;
-  }
-
-  /** Getter for item data context object */
-  get dataContext(): any {
-    return this.args.item;
   }
 
   get editorOptions(): MultipleSelectOption {
@@ -211,10 +183,6 @@ export class SelectEditor implements Editor {
   /** Getter for the Custom Structure if exist */
   protected get customStructure(): CollectionCustomStructure | undefined {
     return this.columnDef?.editor?.customStructure;
-  }
-
-  get hasAutoCommitEdit(): boolean {
-    return this.gridOptions.autoCommitEdit ?? false;
   }
 
   get msInstance(): MultipleSelectInstance | undefined {
@@ -318,11 +286,6 @@ export class SelectEditor implements Editor {
       }
     }
     return '';
-  }
-
-  /** Get the Validator function, can be passed in Editor property or Column Definition */
-  get validator(): EditorValidator | undefined {
-    return this.columnEditor?.validator ?? this.columnDef?.validator;
   }
 
   init(): void {
@@ -637,17 +600,7 @@ export class SelectEditor implements Editor {
 
   /** when it's a Composite Editor, we'll check if the Editor is editable (by checking onBeforeEditCell) and if not Editable we'll disable the Editor */
   protected applyInputUsabilityState(): void {
-    const activeCell = this.grid.getActiveCell();
-    const isCellEditable = this.grid.onBeforeEditCell
-      .notify({
-        ...activeCell,
-        item: this.dataContext,
-        column: this.args.column,
-        grid: this.grid,
-        target: 'composite',
-        compositeEditorOptions: this.args.compositeEditorOptions,
-      })
-      .getReturnValue();
+    const isCellEditable = super.checkInputUsabilityState();
     this.disable(isCellEditable === false);
   }
